Move short to long

+++ src/main/java/com/epl/onebox/model/ActivityTicketTypeAvailability.java
@@ -1,7 +1,7 @@
 //
-// Este archivo ha sido generado por la arquitectura JavaTM para la implantación de la referencia de enlace (JAXB) XML v2.2.8-b130911.1802 
+// Este archivo ha sido generado por la arquitectura JavaTM para la implantaciï¿½n de la referencia de enlace (JAXB) XML v2.2.8-b130911.1802 
 // Visite <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Todas las modificaciones realizadas en este archivo se perderán si se vuelve a compilar el esquema de origen. 
+// Todas las modificaciones realizadas en este archivo se perderï¿½n si se vuelve a compilar el esquema de origen. 
 // Generado el: 2017.03.02 a las 05:38:41 PM CET 
 //
 
@@ -67,7 +67,7 @@ public class ActivityTicketTypeAvailability {
     @XmlElement(name = "availability-info", namespace = "http://www.oneboxtm.es/ns/availability", required = true)
     protected AvailabilityInfo availabilityInfo;
     @XmlAttribute(name = "id", required = true)
-    protected short id;
+    protected long id;
 
     /**
      * Obtiene el valor de la propiedad name.
@@ -217,7 +217,7 @@ public class ActivityTicketTypeAvailability {
      * Obtiene el valor de la propiedad id.
      * 
      */
-    public short getId() {
+    public long getId() {
         return id;
     }
 
@@ -225,7 +225,7 @@ public class ActivityTicketTypeAvailability {
      * Define el valor de la propiedad id.
      * 
      */
-    public void setId(short value) {
+    public void setId(long value) {
         this.id = value;
     }
